+++ src/dap/dap.c
@@ -12,26 +12,26 @@
 #include "../../include/math/vector3.h"
 
 /* Module-level variables for rate derivative calculations */
-static double previousPitchrate = 0.0;
-static double previousYawrate = 0.0;
-static double rollIntegrator = 0.0;
+static float previousPitchrate = 0.0;
+static float previousYawrate = 0.0;
+static float rollIntegrator = 0.0;
 
 Vector3 DAP_Execute(
-    double phi,
-    double accelerationY,
-    double accelerationZ,
-    double rollRate,
-    double pitchRate,
-    double yawRate,
-    double accelerationYCommand,
-    double accelerationZCommand, 
+    float phi,
+    float accelerationY,
+    float accelerationZ,
+    float rollRate,
+    float pitchRate,
+    float yawRate,
+    float accelerationYCommand,
+    float accelerationZCommand, 
     DAPParameters dapParams, 
-    double timeStep
+    float timeStep
 )
 {
-    double deltar = ComputeDeltaRollCommand(phi, rollRate, rollIntegrator, dapParams, timeStep);
-    double deltap = ComputeDeltaPitchCommand(accelerationY, pitchRate, accelerationYCommand, dapParams, timeStep);
-    double deltay = ComputeDeltaYawCommand(accelerationZ, yawRate, accelerationZCommand, dapParams, timeStep);
+    float deltar = ComputeDeltaRollCommand(phi, rollRate, rollIntegrator, dapParams, timeStep);
+    float deltap = ComputeDeltaPitchCommand(accelerationY, pitchRate, accelerationYCommand, dapParams, timeStep);
+    float deltay = ComputeDeltaYawCommand(accelerationZ, yawRate, accelerationZCommand, dapParams, timeStep);
     
     /* Store roll integrator value for next iteration */
     rollIntegrator = deltar;
@@ -40,11 +40,11 @@ Vector3 DAP_Execute(
     return output;
 }
 
-double ComputeDeltaRollCommand(double phi, double rollRate, double intr, DAPParameters dapParams, double timeStep) 
+float ComputeDeltaRollCommand(float phi, float rollRate, float intr, DAPParameters dapParams, float timeStep) 
 {
-    double phiCommand = 0.0; /* assuming roll command is zero */ 
-    double phiErr = phiCommand - phi;
-    double phiLimit;
+    float phiCommand = 0.0; /* assuming roll command is zero */ 
+    float phiErr = phiCommand - phi;
+    float phiLimit;
     
     /* Apply roll angle limits */
     if (phiErr < dapParams.phiMinimum) 
@@ -61,7 +61,7 @@ double ComputeDeltaRollCommand(double phi, double rollRate, double intr, DAPPara
     }
     
     /* Calculate proportional term */
-    double out = phiLimit - dapParams.Kr_roll * rollRate;
+    float out = phiLimit - dapParams.Kr_roll * rollRate;
     
     /* Update integrator with anti-windup */
     intr += out * timeStep;
@@ -75,44 +75,44 @@ double ComputeDeltaRollCommand(double phi, double rollRate, double intr, DAPPara
     }
     
     /* Calculate final control command */
-    double deltaCommandRoll = dapParams.Kp_roll * out + intr * dapParams.Ki_roll;
+    float deltaCommandRoll = dapParams.Kp_roll * out + intr * dapParams.Ki_roll;
     return deltaCommandRoll;
 }
 
-double ComputeDeltaPitchCommand(double accelerationY, double pitchRate, double accelerationYCommand, DAPParameters dapParams, double timeStep) 
+float ComputeDeltaPitchCommand(float accelerationY, float pitchRate, float accelerationYCommand, DAPParameters dapParams, float timeStep) 
 {
-    double currentPitchrate = pitchRate;
+    float currentPitchrate = pitchRate;
     
     /* Calculate pitch rate derivative */
-    double qDot = (currentPitchrate - previousPitchrate) / timeStep;
+    float qDot = (currentPitchrate - previousPitchrate) / timeStep;
     
     /* Calculate acceleration error */
-    double accErrorPitch = accelerationYCommand - (dapParams.Kr_pitch * pitchRate) - 
+    float accErrorPitch = accelerationYCommand - (dapParams.Kr_pitch * pitchRate) - 
                           ((accelerationY + dapParams.c * qDot) * dapParams.Ka_pitch);
     
     /* Update previous pitch rate for next iteration */
     previousPitchrate = currentPitchrate;
     
     /* Calculate final control command */
-    double deltaCommandPitch = accErrorPitch * dapParams.Ks_pitch;
+    float deltaCommandPitch = accErrorPitch * dapParams.Ks_pitch;
     return deltaCommandPitch;
 }
 
-double ComputeDeltaYawCommand(double accelerationZ, double yawRate, double accelerationZCommand, DAPParameters dapParams, double timeStep) 
+float ComputeDeltaYawCommand(float accelerationZ, float yawRate, float accelerationZCommand, DAPParameters dapParams, float timeStep) 
 {
-    double currentYawrate = yawRate;
+    float currentYawrate = yawRate;
     
     /* Calculate yaw rate derivative */
-    double rDot = (currentYawrate - previousYawrate) / timeStep;
+    float rDot = (currentYawrate - previousYawrate) / timeStep;
     
     /* Calculate acceleration error */
-    double accErrorYaw = accelerationZCommand - (dapParams.Kr_yaw * yawRate) - 
+    float accErrorYaw = accelerationZCommand - (dapParams.Kr_yaw * yawRate) - 
                         ((accelerationZ + dapParams.c * rDot) * dapParams.Ka_yaw);
     
     /* Update previous yaw rate for next iteration */
     previousYawrate = currentYawrate;
     
     /* Calculate final control command */
-    double deltaCommandYaw = accErrorYaw * dapParams.Ks_yaw;
+    float deltaCommandYaw = accErrorYaw * dapParams.Ks_yaw;
     return deltaCommandYaw;
 }
